Display: Move listeners to separate HW file.
Mixing this in with the HLE management funcs was messy.
