+++ Core/Debugger/WebSocket/GPUStatsSubscriber.cpp
@@ -19,6 +19,7 @@
 #include <vector>
 #include "Core/Debugger/WebSocket/GPUStatsSubscriber.h"
 #include "Core/HLE/sceDisplay.h"
+#include "Core/HW/Display.h"
 #include "Core/System.h"
 
 struct CollectedStats {
