+++ UWP/CoreUWP/CoreUWP.vcxproj.filters
@@ -751,6 +751,9 @@
     <ClCompile Include="..\..\Core\HW\Camera.cpp">
       <Filter>HW</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\Core\HW\Display.cpp">
+      <Filter>HW</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\Core\Util\PortManager.cpp">
       <Filter>Util</Filter>
     </ClCompile>
@@ -1731,6 +1734,9 @@
       <Filter>MIPS\ARM64</Filter>
     </ClInclude>
     <ClInclude Include="..\..\Core\HW\Camera.h">
+      <Filter>HW</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Core\HW\Display.h">
       <Filter>HW</Filter>
     </ClInclude>
     <ClInclude Include="..\..\Core\Util\PortManager.h">
